test-requirements.txt
-e git+https://github.com/sigmavirus24/github3.py.git@8e9ca0056b8fed956b66dafb5398757cd8d8bed9#egg=Github3.py
GitPython
python-subunit>=0.0.18 # Apache-2.0/BSD
request
testrepository>=0.0.18 # Apache-2.0/BSD
testscenarios>=0.4 # Apache-2.0/BSD
testtools>=1.4.0 # MIT
flake8
pbr>=1.6 # Apache-2.0
GitPython
